Formatting and Content
More messing around. The biggest changes were to trimming bits of content to prevent wrapping around columns in all sizes, actually writing something for In Living Color, and changing the threshold for the "mobile" vs "desktop" site to be 600px wide. The last thing is an attempt to make all content visible at whatever width, but 600 to 800 still sucks.