Slight Bio Tweaks + ID's
Added ID's to a few of the headers (maybe I'll make them links to themselves as anchors? That seems to be common with blind-accessible sites) and tweaked my Bio to line up better with my resume. Still need to add "Award Winning" to my resume on here, off to the work computer.