Fix for no style on certain page widths [Firefox]
Fixed an issue on Firefox (and possibly other browsers) where, when in-between two styles, the site would have _no_ style. This would cause little flashes of white when going between Desktop, Tablet, and Mobile layouts, and if the user lined it up just perfectly could see the whole site unformatted.