Fix for weird column wrap on Chrome
Google Chrome seems to have a different idea of when an item should wrap to the previous column than Firefox does based on the content in those columns; this change is mostly text updates, and as a result also includes a few grammar and spelling fixes.

Only remaining spot that wraps weirdly is the section on Troubadour, where the logo wraps, but I'd like to replace that eventually with a video.